html/semantics/document-metadata/the-meta-element/meta_charset.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: meta_charset</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/document-metadata.html#the-meta-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        assert_equals(document.charset, "UTF-8", "The document charset should be UTF-8.");
      }, "meta_charset");
    </script>
  </body>
</html>
